Move kernel_info, help_links and debugger from the shell to the kernel or interface.
